SISRP-29512 - Updates remaining CS links. Removes campusSolutionsLinkItemDirective and campusSolutionsLinkDirective Renames campusSolutionsLinkV2Directive as campusSolutionsLinkDirective.
