--- src/assets/templates/directives/campus_solutions_link_item.html
@@ -1,9 +1,0 @@
-<a
-  data-ng-bind="text || link.name"
-  data-ng-disabled="{{disabled}}"
-  data-cc-campus-solutions-link-directive="link.url"
-  data-cc-campus-solutions-link-directive-enabled="{{link.isCsLink}}"
-  data-cc-campus-solutions-link-directive-text="link.backToText"
-  data-cc-campus-solutions-link-directive-cache="cache"
-  data-cc-outbound-enabled="{{!link.isCsLink}}">
-></a>
