+++ src/assets/templates/faculty_resources.html
@@ -7,9 +7,9 @@
     <h3>Action Center</h3>
     <ul class="cc-list-bullets">
       <li data-ng-if="facultyResources.eformsReviewCenterLink.url">
-        <a data-cc-campus-solutions-link-v2-directive="facultyResources.eformsReviewCenterLink"
-          data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-          data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
+        <a data-cc-campus-solutions-link-directive="facultyResources.eformsReviewCenterLink"
+          data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+          data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
         ></a>
       </li>
     </ul>
