+++ src/assets/templates/widgets/academics/academic_records.html
@@ -9,9 +9,9 @@
           <a data-ng-href="{{lawTranscriptLink.link}}" data-ng-attr-title="{{lawTranscriptLink.title}}">Transcript - Law School students</a>
         </li>
         <li data-ng-if="lawUnofficialTranscriptLink.url">
-          <a data-cc-campus-solutions-link-v2-directive="lawUnofficialTranscriptLink"
-            data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-            data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
+          <a data-cc-campus-solutions-link-directive="lawUnofficialTranscriptLink"
+            data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+            data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
           ></a>
         </li>
         <li>
