+++ src/assets/templates/widgets/academics/enrollment_verification.html
@@ -44,9 +44,9 @@
                 <strong><a data-ng-href="{{enrollmentGoogleLink.url}}" data-ng-attr-title="{{enrollmentGoogleLink.title}}">Request Official Verification</a></strong>
               </div>
               <strong data-ng-if="requestOfficialVerificationLink.url">
-                <a data-cc-campus-solutions-link-v2-directive="requestOfficialVerificationLink"
-                  data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-                  data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
+                <a data-cc-campus-solutions-link-directive="requestOfficialVerificationLink"
+                  data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+                  data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
                 ></a>
               </strong>
             </div>
