+++ src/assets/templates/widgets/activity_archive.html
@@ -1,10 +1,10 @@
 <div data-ng-if="archiveUrl && (api.user.profile.roles.student || api.user.profile.roles.applicant)" class="cc-text-center cc-widget-padding">
-  <a data-cc-campus-solutions-link-v2-directive
-    data-cc-campus-solutions-link-v2-directive-url="archiveUrl"
-    data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-    data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
-    data-cc-campus-solutions-link-v2-directive-uc-from="true"
-    data-cc-campus-solutions-link-v2-directive-uc-from-link="true"
-    data-cc-campus-solutions-link-v2-directive-uc-from-text="true"
+  <a data-cc-campus-solutions-link-directive
+    data-cc-campus-solutions-link-directive-url="archiveUrl"
+    data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+    data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
+    data-cc-campus-solutions-link-directive-uc-from="true"
+    data-cc-campus-solutions-link-directive-uc-from-link="true"
+    data-cc-campus-solutions-link-directive-uc-from-text="true"
   >Archive of Official Communications</a>
 </div>
