+++ src/assets/templates/widgets/advisor/resources.html
@@ -7,10 +7,15 @@
       <h3>Favorite Reports</h3>
       <ul class="cc-list-links">
         <li data-ng-repeat="report in ucAdvisingResources.ucAdvisingFavoriteReports">
-          <div
-            data-cc-campus-solutions-link-item-directive
-            data-link="report"
-          ></div>
+          <a data-cc-campus-solutions-link-directive
+            data-cc-campus-solutions-link-directive-name="report.name"
+            data-cc-campus-solutions-link-directive-url="report.url"
+            data-cc-campus-solutions-link-directive-uc-from="true"
+            data-cc-campus-solutions-link-directive-uc-from-link="true"
+            data-cc-campus-solutions-link-directive-uc-from-text="true"
+            data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+            data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
+          ></a>
         </li>
       </ul>
     </div>
