+++ src/assets/templates/widgets/enrollment/adjust_waitlisted.html
@@ -7,16 +7,16 @@
       data-ng-if="api.user.profile.actAsOptions.canSeeCSLinks && enrollmentInstruction.isClassScheduleAvailable"
       class="cc-flex cc-flex-wrap cc-enrollment-card-enrolled-links"
     >
-      <a data-cc-campus-solutions-link-v2-directive
-        data-cc-campus-solutions-link-v2-directive-name="'Drop'"
-        data-cc-campus-solutions-link-v2-directive-title="enrollmentInstruction.links.dropEnrolledClasses.name"
-        data-cc-campus-solutions-link-v2-directive-url="enrollmentInstruction.links.dropEnrolledClasses.url"
-        data-cc-campus-solutions-link-v2-directive-cc-cache="enrollment"
-        data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-        data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
-        data-cc-campus-solutions-link-v2-directive-uc-from="true"
-        data-cc-campus-solutions-link-v2-directive-uc-from-link="true"
-        data-cc-campus-solutions-link-v2-directive-uc-from-text="true"
+      <a data-cc-campus-solutions-link-directive
+        data-cc-campus-solutions-link-directive-name="'Drop'"
+        data-cc-campus-solutions-link-directive-title="enrollmentInstruction.links.dropEnrolledClasses.name"
+        data-cc-campus-solutions-link-directive-url="enrollmentInstruction.links.dropEnrolledClasses.url"
+        data-cc-campus-solutions-link-directive-cc-cache="enrollment"
+        data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+        data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
+        data-cc-campus-solutions-link-directive-uc-from="true"
+        data-cc-campus-solutions-link-directive-uc-from-link="true"
+        data-cc-campus-solutions-link-directive-uc-from-text="true"
       ></a>
     </div>
   </div>
