+++ src/assets/templates/widgets/enrollment/decide.html
@@ -9,16 +9,16 @@
 <div class="cc-enrollment-card-section-content" data-ng-if="section.show">
   <div data-ng-if="api.user.profile.actAsOptions.canSeeCSLinks && !isInstructionType(enrollmentInstruction, ['law'])">
     <div data-ng-if="enrollmentInstruction.isClassScheduleAvailable">
-      <a data-cc-campus-solutions-link-v2-directive
-        data-cc-campus-solutions-link-v2-directive-url="enrollmentInstruction.links.chooseClasses.url"
-        data-cc-campus-solutions-link-v2-directive-name="enrollmentInstruction.links.chooseClasses.name"
-        data-cc-campus-solutions-link-v2-directive-title="'Select classes for the upcoming semester.'"
-        data-cc-campus-solutions-link-v2-directive-uc-from="true"
-        data-cc-campus-solutions-link-v2-directive-uc-from-link="true"
-        data-cc-campus-solutions-link-v2-directive-uc-from-text="true"
-        data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-        data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
-        data-cc-campus-solutions-link-v2-directive-cc-cache="enrollment"
+      <a data-cc-campus-solutions-link-directive
+        data-cc-campus-solutions-link-directive-url="enrollmentInstruction.links.chooseClasses.url"
+        data-cc-campus-solutions-link-directive-name="enrollmentInstruction.links.chooseClasses.name"
+        data-cc-campus-solutions-link-directive-title="'Select classes for the upcoming semester.'"
+        data-cc-campus-solutions-link-directive-uc-from="true"
+        data-cc-campus-solutions-link-directive-uc-from-link="true"
+        data-cc-campus-solutions-link-directive-uc-from-text="true"
+        data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+        data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
+        data-cc-campus-solutions-link-directive-cc-cache="enrollment"
       ></a> for the upcoming semester.<br>
     </div>
   </div>
