+++ src/assets/templates/widgets/enrollment/decide_concurrent.html
@@ -9,9 +9,9 @@
     <li>
       <div>
         <div>
-          <a data-cc-campus-solutions-link-v2-directive="enrollmentInstruction.csLinks.ucAddClassEnrollment"
-            data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-            data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
+          <a data-cc-campus-solutions-link-directive="enrollmentInstruction.csLinks.ucAddClassEnrollment"
+            data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+            data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
             data-ng-disabled="enrollmentInstruction.isGradeBaseAvailable || !enrollmentInstruction.csLinks.ucAddClassEnrollment"
           ></a>
         </div>
@@ -27,14 +27,14 @@
       </div>
     </li>
     <li data-ng-if="enrollmentInstruction.csLinks.ucViewClassEnrollment.url">
-      <a data-cc-campus-solutions-link-v2-directive="enrollmentInstruction.csLinks.ucViewClassEnrollment"
-        data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-        data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
+      <a data-cc-campus-solutions-link-directive="enrollmentInstruction.csLinks.ucViewClassEnrollment"
+        data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+        data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
       ></a>
     <li data-ng-if="enrollmentInstruction.csLinks.ucEditClassEnrollment.url">
-      <a data-cc-campus-solutions-link-v2-directive="enrollmentInstruction.csLinks.ucEditClassEnrollment"
-        data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-        data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
+      <a data-cc-campus-solutions-link-directive="enrollmentInstruction.csLinks.ucEditClassEnrollment"
+        data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+        data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
       ></a>
     </li>
   </ul>
