+++ src/assets/templates/widgets/enrollment/explore.html
@@ -50,15 +50,15 @@
       <li data-ng-if="enrollmentInstruction.isClassScheduleAvailable">
         <div>
           <a
-            data-cc-campus-solutions-link-v2-directive
-            data-cc-campus-solutions-link-v2-directive-name="enrollmentInstruction.links.scheduleClasses.name"
-            data-cc-campus-solutions-link-v2-directive-url="enrollmentInstruction.links.scheduleClasses.url"
-            data-cc-campus-solutions-link-v2-directive-cc-cache="enrollment"
-            data-cc-campus-solutions-link-v2-directive-uc-from="true"
-            data-cc-campus-solutions-link-v2-directive-uc-from-link="true"
-            data-cc-campus-solutions-link-v2-directive-uc-from-text="true"
-            data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-            data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
+            data-cc-campus-solutions-link-directive
+            data-cc-campus-solutions-link-directive-name="enrollmentInstruction.links.scheduleClasses.name"
+            data-cc-campus-solutions-link-directive-url="enrollmentInstruction.links.scheduleClasses.url"
+            data-cc-campus-solutions-link-directive-cc-cache="enrollment"
+            data-cc-campus-solutions-link-directive-uc-from="true"
+            data-cc-campus-solutions-link-directive-uc-from-link="true"
+            data-cc-campus-solutions-link-directive-uc-from-text="true"
+            data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+            data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
           >Search for classes</a>
         </div>
       </li>
