+++ src/assets/templates/widgets/enrollment/plan.html
@@ -14,15 +14,15 @@
     <div data-ng-if="enrollmentInstruction.updatePlannerLink && !enrollmentInstruction.academicPlanner.classes.length">
       Build your multi-year
       <a
-        data-cc-campus-solutions-link-v2-directive
-        data-cc-campus-solutions-link-v2-directive-name="enrollmentInstruction.updatePlannerLink.name"
-        data-cc-campus-solutions-link-v2-directive-url="enrollmentInstruction.updatePlannerLink.url"
-        data-cc-campus-solutions-link-v2-directive-cc-cache="enrollment"
-        data-cc-campus-solutions-link-v2-directive-uc-from="true"
-        data-cc-campus-solutions-link-v2-directive-uc-from-link="true"
-        data-cc-campus-solutions-link-v2-directive-uc-from-text="true"
-        data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-        data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
+        data-cc-campus-solutions-link-directive
+        data-cc-campus-solutions-link-directive-name="enrollmentInstruction.updatePlannerLink.name"
+        data-cc-campus-solutions-link-directive-url="enrollmentInstruction.updatePlannerLink.url"
+        data-cc-campus-solutions-link-directive-cc-cache="enrollment"
+        data-cc-campus-solutions-link-directive-uc-from="true"
+        data-cc-campus-solutions-link-directive-uc-from-link="true"
+        data-cc-campus-solutions-link-directive-uc-from-text="true"
+        data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+        data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
       >academic plan</a>
     </div>
 
@@ -33,15 +33,15 @@
         </div>
         <div>
           <a
-            data-cc-campus-solutions-link-v2-directive
-            data-cc-campus-solutions-link-v2-directive-name="enrollmentInstruction.updatePlannerLink.name"
-            data-cc-campus-solutions-link-v2-directive-url="enrollmentInstruction.updatePlannerLink.url"
-            data-cc-campus-solutions-link-v2-directive-cc-cache="enrollment"
-            data-cc-campus-solutions-link-v2-directive-uc-from="true"
-            data-cc-campus-solutions-link-v2-directive-uc-from-link="true"
-            data-cc-campus-solutions-link-v2-directive-uc-from-text="true"
-            data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-            data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
+            data-cc-campus-solutions-link-directive
+            data-cc-campus-solutions-link-directive-name="enrollmentInstruction.updatePlannerLink.name"
+            data-cc-campus-solutions-link-directive-url="enrollmentInstruction.updatePlannerLink.url"
+            data-cc-campus-solutions-link-directive-cc-cache="enrollment"
+            data-cc-campus-solutions-link-directive-uc-from="true"
+            data-cc-campus-solutions-link-directive-uc-from-link="true"
+            data-cc-campus-solutions-link-directive-uc-from-text="true"
+            data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+            data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
           >Update</a>
         </div>
       </div>
