+++ src/assets/templates/widgets/finaid_awards_section.html
@@ -5,12 +5,15 @@
     </div>
     <div>
       <a data-ng-repeat="link in section.links"
-        data-cc-campus-solutions-link-directive="link.url"
-        data-cc-campus-solutions-link-directive-cache="'finaid'"
-        data-cc-campus-solutions-link-directive-enabled="{{link.isCsLink}}"
-        data-cc-campus-solutions-link-directive-text="currentPage.name"
-        data-cc-outbound-enabled="{{!link.isCsLink}}"
-        data-ng-bind="link.title"
+        data-cc-campus-solutions-link-directive
+        data-cc-campus-solutions-link-directive-name="link.title"
+        data-cc-campus-solutions-link-directive-url="link.url"
+        data-cc-campus-solutions-link-directive-cc-cache="finaid"
+        data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+        data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
+        data-cc-campus-solutions-link-directive-uc-from="true"
+        data-cc-campus-solutions-link-directive-uc-from-link="true"
+        data-cc-campus-solutions-link-directive-uc-from-text="true"
         data-ng-class="{'cc-widget-finaid-awards-section-link-separator': ($index > 0)}"
         data-ng-if="api.user.profile.actAsOptions.canSeeCSLinks"
       >
@@ -53,14 +56,17 @@
               </div>
             </div>
             <div data-ng-if="item.rightColumn.link.url" class="cc-widget-finaid-awards-section-item-action">
-              <a
-                class="cc-button cc-button-blue"
-                data-cc-campus-solutions-link-directive="item.rightColumn.link.url"
-                data-cc-campus-solutions-link-directive-cache="'finaid'"
-                data-cc-campus-solutions-link-directive-enabled="{{item.rightColumn.link.isCsLink}}"
-                data-cc-campus-solutions-link-directive-text="currentPage.name"
+              <a class="cc-button cc-button-blue"
+                data-cc-campus-solutions-link-directive
+                data-cc-campus-solutions-link-directive-name="item.rightColumn.link.title"
+                data-cc-campus-solutions-link-directive-url="item.rightColumn.link.url"
+                data-cc-campus-solutions-link-directive-cc-cache="finaid"
+                data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+                data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
+                data-cc-campus-solutions-link-directive-uc-from="true"
+                data-cc-campus-solutions-link-directive-uc-from-link="true"
+                data-cc-campus-solutions-link-directive-uc-from-text="true"
                 data-cc-outbound-enabled="{{!item.rightColumn.link.isCsLink}}"
-                data-ng-bind="item.rightColumn.link.title"
                 data-ng-if="api.user.profile.actAsOptions.canSeeCSLinks"
               ></a>
             </div>
