+++ src/assets/templates/widgets/finaid_communications_tasks.html
@@ -37,14 +37,14 @@
     </li>
     <li data-ng-if="api.user.profile.actAsOptions.canSeeCSLinks">
       <strong>Optional</strong>
-      <a data-cc-campus-solutions-link-v2-directive
-        data-cc-campus-solutions-link-v2-directive-url="communicationsInfo.unsolicitedResourcesUrl"
-        data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-        data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
-        data-cc-campus-solutions-link-v2-directive-uc-from="true"
-        data-cc-campus-solutions-link-v2-directive-uc-from-link="true"
-        data-cc-campus-solutions-link-v2-directive-uc-from-text="true"
-        data-cc-campus-solutions-link-v2-directive-cc-cache="finaid"
+      <a data-cc-campus-solutions-link-directive
+        data-cc-campus-solutions-link-directive-url="communicationsInfo.unsolicitedResourcesUrl"
+        data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+        data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
+        data-cc-campus-solutions-link-directive-uc-from="true"
+        data-cc-campus-solutions-link-directive-uc-from-link="true"
+        data-cc-campus-solutions-link-directive-uc-from-text="true"
+        data-cc-campus-solutions-link-directive-cc-cache="finaid"
       >Upload</a> an Optional Financial Aid Document.
     </li>
   </ul>
