+++ src/assets/templates/widgets/finaid_profile.html
@@ -28,11 +28,14 @@
                   <span class="cc-visuallyhidden" data-ng-bind="item.title"></span>
                 </a>
                 <a data-ng-if="item.link.url && api.user.profile.actAsOptions.canSeeCSLinks"
-                   data-cc-campus-solutions-link-directive="item.link.url"
-                   data-cc-campus-solutions-link-directive-cache="'finaid'"
-                   data-cc-campus-solutions-link-directive-enabled="{{item.link.isCsLink}}"
-                   data-cc-campus-solutions-link-directive-text="currentPage.name"
-                   data-cc-outbound-enabled="{{!item.link.isCsLink}}"
+                  data-cc-campus-solutions-link-directive
+                  data-cc-campus-solutions-link-directive-url="item.link.url"
+                  data-cc-campus-solutions-link-directive-cc-cache="finaid"
+                  data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+                  data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
+                  data-cc-campus-solutions-link-directive-uc-from="true"
+                  data-cc-campus-solutions-link-directive-uc-from-link="true"
+                  data-cc-campus-solutions-link-directive-uc-from-text="true"
                 >
                   <span data-ng-bind="item.link.title"></span>
                   <span class="cc-visuallyhidden" data-ng-bind="category.title"></span>
