+++ src/assets/templates/widgets/profile/delegate.html
@@ -12,13 +12,13 @@
     those privileges at any time.
   </p>
   <p data-ng-if="delegatedAccessUrl.url">
-    <a data-cc-campus-solutions-link-v2-directive
-      data-cc-campus-solutions-link-v2-directive-url="delegatedAccessUrl.url"
-      data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-      data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
-      data-cc-campus-solutions-link-v2-directive-uc-from="true"
-      data-cc-campus-solutions-link-v2-directive-uc-from-link="true"
-      data-cc-campus-solutions-link-v2-directive-uc-from-text="true"
+    <a data-cc-campus-solutions-link-directive
+      data-cc-campus-solutions-link-directive-url="delegatedAccessUrl.url"
+      data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+      data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
+      data-cc-campus-solutions-link-directive-uc-from="true"
+      data-cc-campus-solutions-link-directive-uc-from-link="true"
+      data-cc-campus-solutions-link-directive-uc-from-text="true"
     >Create or manage delegates</a>
   </p>
   <h3>Other information:</h3>
