+++ src/assets/templates/widgets/toolbox/user_preview.html
@@ -92,53 +92,53 @@
               </th>
               <td>
                 <div data-ng-if="ucAdvisingResources.csLinks.studentServiceIndicators.url">
-                  <a data-cc-campus-solutions-link-v2-directive="ucAdvisingResources.csLinks.studentServiceIndicators"
-                    data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-                    data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
+                  <a data-cc-campus-solutions-link-directive="ucAdvisingResources.csLinks.studentServiceIndicators"
+                    data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+                    data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
                   ></a>
                 </div>
                 <div data-ng-if="ucAdvisingResources.csLinks.studentAdvisingAssignments.url">
-                  <a data-cc-campus-solutions-link-v2-directive="ucAdvisingResources.csLinks.studentAdvisingAssignments"
-                    data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-                    data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
+                  <a data-cc-campus-solutions-link-directive="ucAdvisingResources.csLinks.studentAdvisingAssignments"
+                    data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+                    data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
                   ></a>
                 </div>
                 <div data-ng-if="ucAdvisingResources.csLinks.studentAdministrativeTranscripts.url">
-                  <a data-cc-campus-solutions-link-v2-directive="ucAdvisingResources.csLinks.studentAdministrativeTranscripts"
-                    data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-                    data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
+                  <a data-cc-campus-solutions-link-directive="ucAdvisingResources.csLinks.studentAdministrativeTranscripts"
+                    data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+                    data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
                   ></a>
                 </div>
                 <div data-ng-if="ucAdvisingResources.csLinks.studentAcademicProgressReport.url">
-                  <a data-cc-campus-solutions-link-v2-directive="ucAdvisingResources.csLinks.studentAcademicProgressReport"
-                    data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-                    data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
+                  <a data-cc-campus-solutions-link-directive="ucAdvisingResources.csLinks.studentAcademicProgressReport"
+                    data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+                    data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
                   ></a>
                 </div>
                 <div data-ng-if="ucAdvisingResources.csLinks.studentWhatIfReport.url">
-                  <a data-cc-campus-solutions-link-v2-directive="ucAdvisingResources.csLinks.studentWhatIfReport"
-                    data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-                    data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
+                  <a data-cc-campus-solutions-link-directive="ucAdvisingResources.csLinks.studentWhatIfReport"
+                    data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+                    data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
                   ></a>
                 </div>
                 <div data-ng-if="ucAdvisingResources.csLinks.studentManageMilestones.url">
-                  <a data-cc-campus-solutions-link-v2-directive="ucAdvisingResources.csLinks.studentManageMilestones"
-                    data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-                    data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
+                  <a data-cc-campus-solutions-link-directive="ucAdvisingResources.csLinks.studentManageMilestones"
+                    data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+                    data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
                   ></a>
                 </div>
                 <div data-ng-if="ucAdvisingResources.csLinks.studentMultiYearAcademicPlanner.url">
-                  <a data-cc-campus-solutions-link-v2-directive="ucAdvisingResources.csLinks.studentMultiYearAcademicPlanner"
-                    data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-                    data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
-                    data-cc-campus-solutions-link-v2-directive-cc-cache="advisingAcademics"
+                  <a data-cc-campus-solutions-link-directive="ucAdvisingResources.csLinks.studentMultiYearAcademicPlanner"
+                    data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+                    data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
+                    data-cc-campus-solutions-link-directive-cc-cache="advisingAcademics"
                     data-ng-click="expireAcademicsCache()"
                   ></a>
                 </div>
                 <div data-ng-if="ucAdvisingResources.csLinks.studentAdvisorNotes.url">
-                  <a data-cc-campus-solutions-link-v2-directive="ucAdvisingResources.csLinks.studentAdvisorNotes"
-                    data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-                    data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
+                  <a data-cc-campus-solutions-link-directive="ucAdvisingResources.csLinks.studentAdvisorNotes"
+                    data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+                    data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
                   ></a>
                 </div>
                 <div data-ng-if="schedulePlanner">
@@ -150,9 +150,9 @@
                   </a>
                 </div>
                 <div data-ng-if="ucAdvisingResources.csLinks.studentWebnowDocuments.url">
-                  <a data-cc-campus-solutions-link-v2-directive="ucAdvisingResources.csLinks.studentWebnowDocuments"
-                    data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-                    data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
+                  <a data-cc-campus-solutions-link-directive="ucAdvisingResources.csLinks.studentWebnowDocuments"
+                    data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+                    data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
                   ></a>
                 </div>
               </td>
