+++ src/assets/templates/widgets/transfer_credit.html
@@ -79,9 +79,9 @@
           <tr class="cc-transfer-credit-table-parent-row">
             <td colspan="3">
               <a
-                data-cc-campus-solutions-link-v2-directive="transferCredit.tcReportLink"
-                data-cc-campus-solutions-link-v2-directive-cc-page-name="currentPage.name"
-                data-cc-campus-solutions-link-v2-directive-cc-page-url="currentPage.url"
+                data-cc-campus-solutions-link-directive="transferCredit.tcReportLink"
+                data-cc-campus-solutions-link-directive-cc-page-name="currentPage.name"
+                data-cc-campus-solutions-link-directive-cc-page-url="currentPage.url"
               ></a>
             </td>
           </tr>
